added handlers.go for handling requests
